Simplify fixture creation and reduce code duplication in conftest.py
