--- tests/integration/procedure_surface/arrow/conftest.py
@@ -1,45 +1,0 @@
-import logging
-from pathlib import Path
-from typing import Generator
-
-import pytest
-from testcontainers.core.network import Network
-
-from graphdatascience.arrow_client.authenticated_flight_client import AuthenticatedArrowClient
-from graphdatascience.query_runner import QueryRunner
-from graphdatascience.session.dbms_connection_info import DbmsConnectionInfo
-from tests.integration.conftest import (
-    GdsSessionConnectionInfo,
-    create_arrow_client,
-    start_session,
-)
-from tests.integration.procedure_surface.conftest import (
-    create_db_query_runner,
-    start_database,
-)
-
-LOGGER = logging.getLogger(__name__)
-
-
-@pytest.fixture(scope="package")
-def session_connection(
-    network: Network, tmp_path_factory: pytest.TempPathFactory, logs_dir: Path, request: pytest.FixtureRequest
-) -> Generator[GdsSessionConnectionInfo, None, None]:
-    yield from start_session(logs_dir, tmp_path_factory, network, request)
-
-
-@pytest.fixture(scope="package")
-def arrow_client(session_connection: GdsSessionConnectionInfo) -> AuthenticatedArrowClient:
-    return create_arrow_client(session_connection)
-
-
-@pytest.fixture(scope="package")
-def neo4j_connection(
-    network: Network, logs_dir: Path, request: pytest.FixtureRequest
-) -> Generator[DbmsConnectionInfo, None, None]:
-    yield from start_database(logs_dir, network, request)
-
-
-@pytest.fixture(scope="package")
-def query_runner(neo4j_connection: DbmsConnectionInfo) -> Generator[QueryRunner, None, None]:
-    yield from create_db_query_runner(neo4j_connection)
